requirements.txt
ipdb==0.13.9
albumentations==1.0.0
streamlit==0.82.0
tornado==6.1
PyMySQL==0.9.3
scipy==1.7.1
seaborn==0.9.0
sentry-sdk==0.11.1
Shapely==1.8a3
python-dotenv==0.17.1
jupyterlab==3.2.5
pytorch-lightning==1.5.6
wandb==0.10.30
nni==2.8
hydra-core==1.1.1
omegaconf==2.1.1
wget==3.2
yolort==0.6.3
